backend/math-tutor-service/requirements.txt
fastapi==0.115.7
transformers==4.48.1
peft==0.14.0
unsloth
unsloth_zoo
bitsandbytes==0.45.1
uvicorn==0.34.0
pydantic==2.8.2
motor==3.7.0
python-dotenv==1.0.1
llama-cpp-python
redis==5.2.1
